telegram_pre/requirements.txt
telethon
cryptg
hachoir
peewee
sqlalchemy
coloredlogs
python_socks
pyyaml
aiohttp[speedups]
PyMySQL
APScheduler
ton



